types-modified

+++ app/api/products/[id]/route.ts
@@ -1,11 +1,15 @@
 import { NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
 
-export async function GET(
-   request: Request,
-   { params }: { params: { id: string } }
-) {
-   const id = parseInt(params.id);
+// GET product by ID
+export async function GET(request: Request) {
+   const { searchParams } = new URL(request.url);
+   const id = parseInt(searchParams.get("id") as string);
+
+   if (isNaN(id)) {
+      return NextResponse.json({ error: "Invalid ID format" }, { status: 400 });
+   }
+
    const product = await prisma.product.findUnique({
       where: { id },
    });
@@ -17,11 +21,15 @@ export async function GET(
    return NextResponse.json(product);
 }
 
-export async function PUT(
-   request: Request,
-   { params }: { params: { id: string } }
-) {
-   const id = parseInt(params.id);
+// Update product by ID
+export async function PUT(request: Request) {
+   const { searchParams } = new URL(request.url);
+   const id = parseInt(searchParams.get("id") as string);
+
+   if (isNaN(id)) {
+      return NextResponse.json({ error: "Invalid ID format" }, { status: 400 });
+   }
+
    const json = await request.json();
 
    const updatedProduct = await prisma.product.update({
@@ -32,11 +40,14 @@ export async function PUT(
    return NextResponse.json(updatedProduct);
 }
 
-export async function DELETE(
-   request: Request,
-   { params }: { params: { id: string } }
-) {
-   const id = parseInt(params.id);
+// Delete product by ID
+export async function DELETE(request: Request) {
+   const { searchParams } = new URL(request.url);
+   const id = parseInt(searchParams.get("id") as string);
+
+   if (isNaN(id)) {
+      return NextResponse.json({ error: "Invalid ID format" }, { status: 400 });
+   }
 
    await prisma.product.delete({
       where: { id },
